Moved service to addon folder, to make it easier to extend in apps.
